+++ docs/docs/sandpack/files/media/code-imagePlugins.ts
@@ -1,10 +1,13 @@
 export const imagePluginsCode = `import {
   createBasicElementsPlugin,
+  createComboboxPlugin,
+  createEmojiPlugin,
   createImagePlugin,
   createSelectOnBackspacePlugin,
 } from '@udecode/plate';
 import { basicMarksPlugins } from '../basic-marks/basicMarksPlugins';
 import { plateUI } from '../common/plateUI';
+import { emojiPlugin } from '../emoji/emojiPlugin';
 import { selectOnBackspacePlugin } from '../select-on-backspace/selectOnBackspacePlugin';
 import { createMyPlugins } from '../typescript/plateTypes';
 
@@ -14,6 +17,8 @@ export const imagePlugins = createMyPlugins(
     ...basicMarksPlugins,
     createImagePlugin(),
     createSelectOnBackspacePlugin(selectOnBackspacePlugin),
+    createComboboxPlugin(),
+    createEmojiPlugin(emojiPlugin),
   ],
   {
     components: plateUI,