+++ examples/apps/next/src/config/sidebarItems.ts
@@ -8,6 +8,7 @@ export const sidebarItems = [
   'cursor-overlay',
   'dnd',
   'editable-voids',
+  'emoji',
   'excalidraw',
   'exit-break',
   'find-replace',