+++ packages/headless/src/index.tsx
@@ -12,6 +12,7 @@ export * from '@udecode/plate-button';
 export * from '@udecode/plate-code-block';
 export * from '@udecode/plate-combobox';
 export * from '@udecode/plate-core';
+export * from '@udecode/plate-emoji';
 export * from '@udecode/plate-find-replace';
 export * from '@udecode/plate-floating';
 export * from '@udecode/plate-font';