+++ packages/ui/plate/src/index.tsx
@@ -9,6 +9,7 @@ export * from '@udecode/plate-ui-code-block';
 export * from '@udecode/plate-ui-combobox';
 export * from '@udecode/plate-ui-cursor';
 export * from '@udecode/plate-ui-dnd';
+export * from '@udecode/plate-ui-emoji';
 export * from '@udecode/plate-ui-find-replace';
 export * from '@udecode/plate-ui-font';
 export * from '@udecode/plate-ui-horizontal-rule';